osd/: fast dispatch, standard op path for MOSDRepScrub
Previously, it had its own queue, but better to simply use
the OpWQ along with everything else.  It also allows MOSDRepScrub
to be fast dispatched.  We use NORMAL priority since we want the
replica scrub message processed before recovery or client IO
messages -- the primary is blocking writes on the objects until
the replica responds.

Signed-off-by: Samuel Just <sjust@redhat.com>